0c3126f remove test, it's not worth the build server hassle
3d61ae6 Ninja: remove int/size_t warning

+++ Source/cmGlobalNinjaGenerator.cxx
@@ -558,7 +558,7 @@ void cmGlobalNinjaGenerator::AddRule(const std::string& name,
                                     restat,
                                     generator);
 
-  this->RuleCmdLength[name] = command.size();
+  this->RuleCmdLength[name] = (int) command.size();
 }
 
 bool cmGlobalNinjaGenerator::HasRule(const std::string &name)
